+++ src/ctmysql/__init__.py
@@ -1,0 +1,6 @@
+"""关系数据库MySQL版(RDS)管理模块"""
+
+from .client import RDSClient
+from .commands import ctmysql
+
+__all__ = ['RDSClient', 'ctmysql']